Officially added the tcp_client_permanent_message_connection.cpp.
Deleted the "communicator_dispatcher.cpp" from the CMakeLists.txt, it doesn't exist.
Added all the header files to the list of headers to install.
Added a \file comment in all the files so they will appear in the docs.
Fixed copyright notices.
Fixed #include's (position, "" and <>, corresponding #define).
Removed an obsolete function.
Removed an if() which is not necessary.
Tweaked the Doxy file to remove the full path to files. It's not sensible in the docs.