Added the Qt connection as a separate library.
Added Qt5 as a dependency.
Fixed the use of the SNAP_LOG to work with the new logger in the dispatcher.
Fixed some namespace references in the dispatcher.
Fixed the `<<` which worked with `QStringList` but not `string_list_t`.
Fixed includes for the dispatcher.
Fixed various comments still referencing snap this, snap that.
Use nproc in the mk script.
Ignore swp files.