Added a signal_profiler to capture SIGPROF signals.
When using the profiler against a tool that uses any system calls, you may
get a SIGPROF signal and then your system call returns with EINTR. That
means you have to call the function again. Only that's very tedious. So
this class is expected to prevent that requirement. At this time, it doesn't
work yet, but I'm always hopeful.