Print out the errno info.
I'm, at times, getting that warning and I then lose the listen() itself
(which sounds really strange/weird). The BIOS may do the close of the
accept socket, but either way, it's a hardship since after such an
event that server can't accept any more connections since it's closed
and I'm not even aware of it!