Updated the signal_handler to make it usable.
First of all, having a virtual callback is not going to work if the object
is a singleton (i.e. you can't do new ... so it won't work at all). So I
replaced that with a list of callbacks with an identifier.

Make use of the signal_handler object in snaploggerd.