Very first step converting the Snap! Communicator.
This code comes from the snapwebsites/snapcommunicator/src folder. The
very first step was to break out all the classes in a separate file.
It won't compile for a while as I decided that the fluid-settings
should work on its own and allow already manage the snapcommunicator
daemon settings (except for the one connection to the fluid-settings).

This will resolve many management of issues of a large cluster.

Added a broadcast_message() in the connection_with_send_message class
so we can easily send a message to many connections as held by some
container.