Added the accept() function to the local_stream_server_connection.
In the TCP implementation, this is part of the tcp_bio_server.
Also added the close-on-exec set/get functions so we can change that flag between the main server and the accepted connections.
Made a note that the tcp_client.cpp/h and tcp_server.cpp/h were not actually used by the communicator.
Added a catch_unix.cpp test to make sure that the AF_UNIX socket works.
Properly initialize the logger in the tests.
Renamed the test file so it starts with catch_...