Added support for a LEAK message.
When compiling with the sanitizer, we have the ability to print out the
list of memory blocks currently allocated. If we have a tool which over
time seems to leak, then this is useful to find out what buffers are
allocated and never released. To detect the leaks, send a LEAK message
to your tool (assuming it accepts the default messages supported by
the dispatcher) and your tool will spit out the leak information in
your console (we want to look at changing the output in a future
release so it gets printed in a leak specific log file).