Added an alert appender.
The idea is to count the number of filtered messages that arrive to this
appender. After a certain number or certain types ("alert" component),
those messages get forwarded via the TCP connection. The daemon on the
other side can then forward those messages to an email address or a
system such as slack.