Added support for direct files.
Added a test to verify that direct files work as expected.
Fixed a but in the add_listener() call, we need to specify the correct object, not always 'this'.
Added a process_done() which gets called when the SIGCHLD of a process is received.
Added an output_done() that gets called with the output once it is complete (capture mode).
Added an error_done() that gets called with the output errors once it is complete.
Added documentation about the cppprocess and equivalence to a shell.
Added an add_connection() & a remove_connection() to signal_child.
Fixed the add/remove connection of the signal_child so both can be done multiple as expected.
Automated the add/remove connection of the signal_child in the add_listener() and process_signal() functions.