cppprocess makes use of I/O objects
The first cppprocess implementation presented here mostly came from the
libsnapwebsites. This one is an improvement (I think) which makes use of
separate objects to define the input and output to use. Either a file or
a pipe at the moment. The file can only be defined from a filename, but
I intend to add a version where you can specify a file descriptor.

Added Boost as a required dependency.
Added a couple of TODOs.
Re-added the in-use exception.
Couple fixes to the copyright notice.
Updated the tests to work with the new scheme (I/O instead of flags & such).