Added a way to auto-list connections.
Often, to debug connections, I list the ones still defined in the communicator
because that's the one that prevent a QUIT from working. This can now be
automated.

Applied a fix to the signal handler, avoid setting the signal is the number
is negative (which is the default in the class).

Fixed the doxy file: the name is just eventdispatcher, the source is the
one line that has to use both eventdispatcher and eventdispathcher_qt.