Fixed up some documentation about cppprocess.
Especially, the `f_child` parameter can be retrived with `p.process_pid()`.
This fix to the docs included the necessary addition of `process::process_pid()`.
Added a function to transform the command line to one string.
Added a function to retrieve the exit code of a process (useful to check intermediate processes).
Reset f_child once it is done (i.e. set back to -1, i.e. the process is not running).
Removed the old function names from the class definitions. Everything else works.
Updated the advgetopt structures.