Added defaults to the daemon command line options.
Updated the snaplogger daemon with the new logrotate options.
Actually send the messages we receive from clients.
Changed the severity to use the name instead of a number.
Actually set the timestamp.
Actually outputs the message itself.
Fixed the name of the snaplogger network plugin.
Renamed the counter field "standard" (as in standard limit for standard events opposed to limit of alerts).
Fixed spelling of the "alert" field.
Reset alert counters back to zero instead of just `--`.
Removed `snaplogger-network` as a library of `snaplog`. It's a plugin.