Many more fixes to the snapcommunicator.
This is an intermediate push. I am actually moving the snapcommunicator to
its own project because (1) it's huge and (2) it requires the edhttp which
manages flag files.

I also fixed a couple of doc/CMakeLists.txt as I moved the Doxygen search
locally.

I added the CUI connection, which is a terminal (curses) connection to handle
CUI through the eventdispatcher.

Moved the snapstop as an eventdispatcher tool named ed-stop. It can be used
to stop a service. This is very similar to the systemctl stop commands at
the moment. I'd like to be able to send the STOP message as well.

Various clean ups.