Added the cui_connection here.
Added a weak_pointer_t type to the ed::connection class.
Added a user pointer to the ed::message class.
Removed the `const` from the send_message(). We remove it in many cases and 99.999% of the time, the caller doesn't care about that const anyway.
Made the dispatch_message() function virtual for the snapcommunicatord.
Moved the mode_t to utils.h and share between the tcp_bio_client/server.
Various clean ups, documentation, logs, etc.