Moved the broadcast function to its own header.
The broadcast function sends to many different connections and thus it was
a static function. Therefore, I moved it out of the
connection_with_send_message class and also I made it more reliable (i.e.
it supports containers with null pointers and either shared or weak pointers).

Renamed the exception without an additional namespace in said name.