Added kludge to get around BIO bug.
The older BIO implementation had an issue where the bind() could fail and the
error would not be reported. I added a _kludge_ to verify that didn't happen.
The fact is the socket number is still available but the socket itself is
closed so we can call getsockopt() on it and it will return an error (EBADF)
when the bind() failed.

I also changed the BIO_do_accept() with BIO_do_connect() because I think it
looks better in the code. At that point, we are not accepting anything. The
fact is that internally, both call BIO_handshake() anyway.

Added a warning to explain why we see a fatal error when creating a Unix
stream and it continues to work! (i.e. it is a test to verify that the socket
is available).

Fixed some includes that used "..." instead of <...>.
Deleted some commented out code.
Removed support for Ubuntu 16.04 (std::clamp() is always present now).
Fixed a few !<ptr> with <ptr> == nullptr.