Made logrotate_extension a virtual.
The logrotate_extension is more than likely to be used in classes which have
one or more virtual functions so I made it have a virtual table by adding
a virtual destructor. That way we avoid the compile time message saying
some class destructor are not virtual (plus this one does require clean up).

Added support for the regular send_message() in the udp_server_message_connection
including a way to define a permanent secret code.