Removed duplicated connection with send message.
Since I added the connection_with_send_message to the
upd_server_message_connection class, we do not need it in the
logrotate_udp_messenger and udp_logger_server. Plus, it was not implemented
on those classes (i.e. it would just throw an error if called). Those two
other classes do not make use of the send_message() at this point, though,
since it does not define a destination.