The UDP & Local connections have a service name.
Like the TCP connection, we can now use the UDP and Local connections (Unix
streams) to connect to the communicator daemon so we need to have support for
a service name. This is now available.

Added a sample for a new version of the event dispatcher which allows us
to define functions so that we do not need to reimplement new classes each
time we want to instantiate a connection. This is for a future version. I'm
not really sure thus far that it would work as presented thus far.