Finally removed the <T> in the dispatcher.
This version of the dispatcher makes do without a template. This allows use to
add functions at any level (i.e. we have a communicator, a fluid-service, and
the user service, for example, we may have even more later).

I broke up the dispatcher_match class to its own set of files. Moved the code
to a .cpp file. Removed the dispatcher_base since it is not required.

Updated all the code that now uses the new scheme. That includes the new
callback in the dispatcher_match, instead of a simpler &T::<name>, we now
use an std::function<void(ed::message&)> allowing for functions to be defined
absolutely anywhere.

Added the show_matches flag so while adding new COMMANDs, it gets printed
to the logs. This is useful to debug the list of commands a service supports.

Fixed the #include to use <...> when it appears in a header.