Renamed the "addr::unix" class as "addr::addr_unix".
The gnu++17 extension defines a macro named "unix" which prevents us from
having such a class. I updated the libaddr library already. Now I updated
the eventdispatcher.

Also, the Qt5X11Extras would append the CMAKE_CXX11_STANDARD_COMPILE_OPTION
to the CXX_FLAGS. As a result, we would have -std=c++11 instead of our
-std=g++17 (because it appears after it has priority). So I now set that
variable to "" to prevent the override. This is ugly, but it looks like it
works as expected.