Fixed the filename.
The filename is given a length, if 0, then use std::string(). If larger
than zero, do not use the length because it has padding at the end and
we'd end up with a string including '\0' characters which we do not want.

Added the UPDATED event to painlessly catch the open+write+close sequence.
Extracted the event definitions out of the file_changed class.
Changed a WARNING to a FATAL error since we throw in that case.
Properly handle case where ievent.wd == -1 (queue full).
Added tests to verify the file_changed class.
Updated some copyright notices.
Various documentation clean ups.