Fixed up some mistakes in the file_changed impl.
Added a note about the idea of support recursivity for file_changed (and
started by at least having a RECURSIVE flag).

Implemented an EXISTS flag to get events for existing files when you first
setup a watch (because otherwise you are not told about those files). Still
need to implement the glob & loop.