Added snaplog to a package.
I created a new package, snaplogger-utils, for the snaplog tool. This tool
requires the eventdispatcher library. It could not be placed in the
snaplogger-network plugin package because we do not want to force people
to install that plugin to use the snaplog tool. Similarly, the snaploggerd
package is very specific to the deamon (listener) of the network packets
and that one is generally installed on a single computer. So again, the
snaplog was not welcome in that package.

I also did some small clean ups (main of headers and documentation).