Fixed memory leaks when running coverage.
The OpenSSL still leaks here and there. Small buffers, but still. So we
need to setup the sanitizer to ignore a certain number of OpenSSL functions
in order for the coverage (which runs with the sanitizer turned on) to work.

Added support for list of parameters in verify_message().

This is quite complete. I will now be working on (close to) full coverage.