Added a "processed" flag to the message class.
When running the dispatcher and at least one CALLBACK is run against a
message, it is now considered processed whether it gets captured later
by another function or not. A processed message does not generate the
UNKNOWN reply automatically. This is useful for some special messages
such as the TRANSMISSION_REPORT as used by the cluck project.