Fixed the kill() instruction.
Now the kill() instruction uses the thread handle and pthread_kill() so we
make sure the signal is sent to that one task instead of a random thread as
selected by the kernel.

Added a warning to the signal implementation when we detect the signal was
already blocked (probably meaning that it is handled somewhere else).

Also there was a bug saving an errno in the wrong place.