Added support for timestamp in verifying messages.
Fixed various documentations.
Updated the ed-signal so the [[<server>:]<service>/]<command> syntax works.
Removed one last IP:port as a string instead of addr::addr.

Cleaned up some of the logrotate classes, but I think it's not that necessary
anymore (better send the info to communicatord and let the rest happen as
expected).