Got 100% of the tests to pass again.
Added type of timestamp param of ABSOLUTELY and ALIVE as a timespec value.
Fixed crash when trying to get location without a running statement.
Prevent crash in the sort() command when the variable does not exist.
Fixed return type of get_label_position(), it's a line number, not an IP.
Added paths to directories with message definitions.