Fixed the initialization (I used the wrong function).
Added a test for the currently unsupported backslashes.
Fixed one exception that used std:: instead of ed::...
Added a test to verify the <=> operator parsing in the lexer.
Added a test to show that unimplemented backslashes are properly detected.
Added a test to show that unsupported escape sequences generate an error.
Actually added the message definitions for the tests to the repository.