Updated the docs & reporter tests.
Changed the reporter exception so all make use of eventdispatcher exceptions.
Fixed the type of the variable names in the sort() parameters. They have to be strings or identifiers.
Allow the "command" of a message to be verified using a regex.
Removed many of the tests in the instructions when those are tested earlier (LCOV_EXCL_LINE is a prime example of why we do not need that code).
Added location to more exceptions.
Allow the signal from the kill() command to either be a number (integer) or a name (string or identifier).
Removed DOUBLE STRING support in a few places where it is already converted to a SINGLE STRING.
Fixed a bug in the hexadecimal parsing, empty (0x) would return 0 instead of an error.
Updated many tests to get 100% coverage and catch the new exceptions.