Added support to retrieve certificates.
This version includes a certificate class that allows us to get a certificate
from a website that has a TLS certificate or loaded from a .pem file. The
certificate names, not before and not after dates is retrieved and made
available to the user. With time, other parameters will be made available.