Finally fixed the leaks!
Before calling the BIO_free_all(), I was closing the socket in order to
avoid the call to the shutdown() function. That shutdown prevents the
system from using the same BIO object in a parent and a child process
after a fork().

Instead I now re-implemented the shutdown() function and do nothing in it.
This is very similar to what I was trying to do earlier, except that now
the BIO interface properly releases all the resources as expected.

The rest is just adjustments to the certificate implementation. Also the
GENERAL_NAMES needs to be freed. That was actually the only leak I've
fixed after the fix to the BIO_free_all() function.