Added support for a callback in the timer class.
This will allow us to create separate objects with a callback to a function
that wants to listen for the ticks. Much easier than the existing version.

Updated a couple of tests that were affected by some changes in the snapdev
library (timespec_ex now uses UTC pretty much everywhere).