Implemented the cache bypass for tcp client connections.
The TCP connection may be able to directly write to the socket and if that
works in full, avoid caching any data. This will work for many smaller
messages which is the greater majority of our traffic.

Fixed the socket validity tests using the valid_socket() function instead
of inferring -1 as an invalid value. Also in one case we checked > 0 instead
of >= 0.

Tweaked the event maximum count to use 8 bits. It's probably more than enough
since we default to 5 and don't expect users to increase that number by much
if at all.