Started work adding support for non-message connections.
The eventdispatcher allows for different types of connections, most of which
can use binary (even if most of our services use the message interface, the
main majority of the classes allow for binary data to travel).

This implements a "direct connection" in our reporter to allow us to debug
binary TCP clients.