Added an idle counter.
When the run() command waits for another message, it is considered idle.
The new version saves the time before calling poll() and the time just
after. It then compiles how long it took and saves that in the idle counter.
At any time, you can get that counter for any purpose such as to know whether
your service is constantly running.

Also fixed the help sentence about the limit in check-certificate.