Updated snaplogger to support fallbacks.
I see that I also had some specialized console fallbacks in the network.
I may want to remove all of those since I now have that properly implemented
(a.k.a. we can now define a list of fallback appenders which are any of
the other appenders instead of specialized code). Next versions.