Added a --process-name option to ed-stop.
The ed-stop command line can be used to send a signal (SIGINT, SIGTERM...)
to a process. By default, it expects a PID. If the value passed is not just
a plain decimal number, then the process searches the list of running
services and if found use that service PID. If not found, you can now use
the --process-name option to allow for the search of the process in the
list of processes (i.e. the same list you get with `ps -ef`). This uses
a non-default setup because it can be dangerous to send a signal such
as SIGINT to a process not listening for that signal.